fix: address final review findings — revert SELECT 1::BIGINT scope creep, standardize boolean casing
- Revert unplanned SELECT 1::BIGINT to SELECT 1 in application_service.rs
- Standardize moderation.rs and push_notification.rs boolean literals to
  uppercase TRUE/FALSE for consistency with the other 13 files in Phase C
- federation_blacklist.rs left as-is (uses sqlx::query! which requires
  cache regeneration to change)

+++ synapse-storage/src/application_service.rs
@@ -736,7 +736,7 @@ impl ApplicationServiceStorage {
     pub async fn has_exclusive_user_namespace_match(&self, as_id: &str, user_id: &str) -> Result<bool, sqlx::Error> {
         let matched = sqlx::query_scalar::<_, i64>(
             r"
-            SELECT 1::BIGINT
+            SELECT 1
             FROM application_service_user_namespaces
             WHERE as_id = $1
               AND is_exclusive = TRUE
